requirements (1).txt
discord.py[voice]
aiosqlite
motor